Struct github_rs::github::Client

pub struct Client {
    pub token: AccessToken,
    pub headers: Headers,
}

Struct used to make calls to the Github API.

Fields
token: AccessToken

Token generated from here on Github that grants the Client access to various parts of the API depending on what is allowed with that token.

headers: Headers

Headers that are automatically defined and used in API Calls to the GithubAPI via the new function. If you wish to define your own to change how you make calls you should modify this variable. However, this is discouraged as the automatically created Headers contain the Authorization, API Version, Media type, and UserAgent headers for you. If you must do so it might cause undefined behavior for the API and errors encountered are not tested for and not the fault of the library authors

Methods

impl Client

fn new(acc_token: AccessToken) -> Result<Client>

Generate a new Client struct with the access token so that calls can be made to the API using it