Trait github_rs::github::Activity

pub trait Activity {
    fn get_events(&self) -> Result<Vec<Event>>;
    fn get_repos_owner_repo_events(&self,
                                   owner: &str,
                                   repo: &str)
                                   -> Result<Vec<Event>>;
    fn get_repos_owner_repo_issues_events(&self,
                                          owner: &str,
                                          repo: &str)
                                          -> Result<Vec<Event>>;
    fn get_networks_owner_repo_events(&self,
                                      owner: &str,
                                      repo: &str)
                                      -> Result<Vec<Event>>;
    fn get_orgs_org_events(&self, org: &str) -> Result<Vec<Event>>;
    fn get_users_username_received_events(&self,
                                          username: &str)
                                          -> Result<Vec<Event>>;
    fn get_users_username_received_events_public(&self,
                                                 username: &str)
                                                 -> Result<Vec<Event>>;
    fn get_users_username_events(&self, username: &str) -> Result<Vec<Event>>;
    fn get_users_username_events_public(&self,
                                        username: &str)
                                        -> Result<Vec<Event>>;
    fn get_users_username_events_orgs_org(&self,
                                          username: &str,
                                          org: &str)
                                          -> Result<Vec<Event>>;
}

Trait used to define access to endpoints grouped under Activity in the Github API specification

Required Methods

fn get_events(&self) -> Result<Vec<Event>>

Request Type:

GET

Endpoint:

/events

Description

Returns a vector of recent Events from the API

fn get_repos_owner_repo_events(&self,
owner: &str,
repo: &str)
-> Result<Vec<Event>>

Request Type:

GET

Endpoint:

/repos/:owner/:repo/events

Description

Returns a vector of Events from the repo of a given owner

fn get_repos_owner_repo_issues_events(&self,
owner: &str,
repo: &str)
-> Result<Vec<Event>>

Request Type:

GET

Endpoint:

/repos/:owner/:repo/issues/events

Description

Returns a vector of Events from the issues of a repo of a given owner

fn get_networks_owner_repo_events(&self,
owner: &str,
repo: &str)
-> Result<Vec<Event>>

Request Type:

GET

Endpoint:

/networks/:owner/:repo/events

Description

Returns a vector of Events from the network of a repo of a given owner

fn get_orgs_org_events(&self, org: &str) -> Result<Vec<Event>>

Request Type:

GET

Endpoint:

/orgs/:org/events

Description

Returns a vector of Events from a given organization

fn get_users_username_received_events(&self,
username: &str)
-> Result<Vec<Event>>

Request Type:

GET

Endpoint:

/users/:username/received_events

Description

Returns a vector of Events the username has received. If you are authenticated and use your own username you will see private Events

fn get_users_username_received_events_public(&self,
username: &str)
-> Result<Vec<Event>>

Request Type:

GET

Endpoint:

/users/:username/received_events/public

Description

Returns a vector of public Events the username has received.

fn get_users_username_events(&self, username: &str) -> Result<Vec<Event>>

Request Type:

GET

Endpoint:

/users/:username/events

Description

Returns a vector of Events performed by the username. If you are authenticated and use your own username you will see private Events

fn get_users_username_events_public(&self, username: &str) -> Result<Vec<Event>>

Request Type:

GET

Endpoint:

/users/:username/events/public

Description

Returns a vector of Events performed by the username. If you are authenticated and use your own username you will see private Events

fn get_users_username_events_orgs_org(&self,
username: &str,
org: &str)
-> Result<Vec<Event>>

Request Type:

GET

Endpoint:

/users/:username/events/orgs/:org

Description

Returns a vector of Events from the username's org. You must be authenticated to view this.