Trait github_rs::github::Repos

pub trait Repos {
    fn post_org_repos(&self,
                      organization: String,
                      repo: RepoCreate)
                      -> Result<Repo>;
    fn post_user_repos(&self, repo: RepoCreate) -> Result<Repo>;
}

Trait definition for endpoints grouped under Repositories in the Github API specification

Required Methods

fn post_org_repos(&self, organization: String, repo: RepoCreate) -> Result<Repo>

Request Type:

POST

Endpoint:

/orgs/:org/repos

Description

Creates a new repo in the specified organization for the authenticated user and returns the new Repos stats

fn post_user_repos(&self, repo: RepoCreate) -> Result<Repo>

Request Type:

POST

Endpoint:

/user/repos

Description

Creates a new repo for the authenticated user and returns the new Repos stats